[cpp] Increase cap for exp/cp bonus.
Changed int16 for cap (subpower) to int32 to allow for higher bonuses to be applied to experience items, or even with using !addeffect.

+++ src/map/utils/charutils.cpp
@@ -4731,7 +4731,7 @@ namespace charutils
         {
             CStatusEffect* commitment = PChar->StatusEffectContainer->GetStatusEffect(EFFECT_COMMITMENT);
             int16          percentage = commitment->GetPower();
-            int16          cap        = commitment->GetSubPower();
+            int32          cap        = commitment->GetSubPower();
             rawBonus += std::clamp<int32>(((capacityPoints * percentage) / 100), 0, cap);
             commitment->SetSubPower(cap -= rawBonus);
 
@@ -5963,7 +5963,7 @@ namespace charutils
         {
             CStatusEffect* dedication = PChar->StatusEffectContainer->GetStatusEffect(EFFECT_DEDICATION);
             int16          percentage = dedication->GetPower();
-            int16          cap        = dedication->GetSubPower();
+            int32          cap        = dedication->GetSubPower();
             bonus += std::clamp<int32>((int32)((exp * percentage) / 100), 0, cap);
             dedication->SetSubPower(cap -= bonus);
